- In the main loop, placed all calls to programming functions inside the check for programmingScope so regular messages are not interpreted as programming messages - copyPage function has unsigned parameter

+++ lithneco/src/lithneProgrammer.cpp
@@ -354,7 +354,7 @@ bool LithneProgrammer::stopProgrammer(void){
 	return success;
 }
 
-bool LithneProgrammer::copyPage( int pageNum )
+bool LithneProgrammer::copyPage( uint16_t pageNum )
 {
 	// We pour blocks of data stored in buffer[], ready to write to the AVR's flash
 	// We only write in blocks of 512 bytes for the atXmega256.
